Chore: faq padding 조정

See merge request sw_track/class_08/web_project_i/team05/frontend!27

+++ public/js/admin/product.js
@@ -1,3 +1,44 @@
+// import { API_URL } from '/public/js/constants.js'
+
+const getProducts = async () => {
+  const token = localStorage.getItem('jwt')
+  const res = await fetch('http://localhost:8081/api/v1/admin/products', {
+    method: 'GET',
+    headers: {
+      Authorization: token,
+    },
+  })
+  if (!res.ok) {
+    throw new Error('서버에서 데이터를 가져오는 데 실패했습니다.')
+  }
+  const productsDatas = await res.json()
+  const adminContainerEl = document.querySelector('.admin-container')
+  productsDatas.data.forEach(data => {
+    adminContainerEl.insertAdjacentHTML(
+      'beforeend',
+      `
+      <div class="admin-product-data">
+          <p class="product-id">${data._id}</p>
+          <p class="product-category">${data.category}</p>
+          <p class="product-title">${data.title}</p>
+          <img src="${data.main_image.url}" class="product-img"></img>
+          <p class="product-price">${data.price}</p>
+          <p class="product-stock">${data.stock}</p>
+          <p class="product-description">${data.description}</p>
+          <p class="product-size">${data.size}</p>
+          <p class="product-origin">${data.origin}</p>
+          <p class="product-attribute">${data.attribute}</p>
+          <div>
+            <button class="btn product-update">상품수정</button>
+            <button class="btn product-delete">상품삭제</button>
+          </div>
+        </div>
+      `
+    )
+  })
+}
+getProducts()
+
 // 상품 추가 버튼 이벤트 리스너
 const productAddBtn = document.querySelector('.product-add')
 if (productAddBtn) {
@@ -6,21 +47,24 @@ if (productAddBtn) {
 
 // 관리자 데이터에서 발생하는 클릭 이벤트 처리
 const adminContainer = document.querySelector('.admin-container')
-adminContainer.addEventListener('click', e => {
-  if (e.target.className.includes('product-cancel')) {
-    handleProductAddCancel(e)
-  } else if (e.target.className.includes('product-create')) {
-    handleCreateBtn(e)
-  } else if (e.target.className.includes('product-update')) {
-    handleUpdateProduct(e)
-  } else if (e.target.className.includes('product-delete')) {
-    handleDeleteProduct(e)
-  } else if (e.target.className.includes('update-cancel')) {
-    handleUpdateCancel(e)
-  } else if (e.target.className.includes('update-save')) {
-    handleUpdateSave(e)
-  }
-})
+if (adminContainer) {
+  adminContainer.addEventListener('click', e => {
+    if (e.target.className.includes('product-cancel')) {
+      handleProductAddCancel(e)
+    } else if (e.target.className.includes('product-create')) {
+      handleCreateBtn(e)
+    } else if (e.target.className.includes('product-update')) {
+      handleUpdateProduct(e)
+    } else if (e.target.className.includes('product-delete')) {
+      handleDeleteProduct(e)
+    } else if (e.target.className.includes('update-cancel')) {
+      handleUpdateCancel(e)
+    } else if (e.target.className.includes('update-save')) {
+      handleUpdateSave(e)
+    }
+  })
+}
+
 function handleCreateProduct() {
   // 상품 추가 UI 생성
   // 상품 생성 UI 핸들러
@@ -39,7 +83,7 @@ function handleCreateProduct() {
         <option value="ACC">ACC</option>
       </select>
       <input / class="product-input title" type='text' placeholder="제품명 입력">
-      <label for="input-img" class="img-label">이미지첨부(5개)</label>
+      <label for="input-img" class="img-label">이미지첨부(6개)</label>
       <input / class="product-input image" id="input-img" type='file' accept='image/*' multiple>
       <input / class="product-input price" type='number' placeholder="가격 입력">
       <input / class="product-input stock" type='number' placeholder="재고 입력">
@@ -91,49 +135,59 @@ function handleCreateBtn(e) {
   const stock = select('.stock')
   const description = select('.description')
   const attribute = select('.attribute')
-  const files = select('.image').files
+  const file = select('.image')
+  const category = select('.category')
+  const size = select('.size')
+  const origin = select('.origin')
 
   // 제품 생성할 때 input 올바르게 작성했는지 검사
   if (
     !title.value ||
+    !title.value.trim() ||
     !price.value ||
-    files.length !== 5 ||
-    !stock ||
+    !price.value.trim() ||
+    file.files.length !== 6 ||
+    !stock.value ||
+    !stock.value.trim() ||
     !description ||
-    !attribute
+    !description.value.trim() ||
+    !attribute.value ||
+    !attribute.value.trim()
   ) {
-    alert('빈칸을 전부 작성하셔야 생성 가능합니다.(이미지 5개)')
+    alert('빈칸을 전부 작성하셔야 생성 가능합니다.(이미지 6개)')
     return
   }
 
   // 데이터 객체 생성
   const datas = {
-    category: select('.category').value,
+    category: category.value,
     title: title.value,
     price: price.value,
     stock: stock.value,
-    size: select('.size').value,
-    origin: select('.origin').value,
+    size: size.value,
+    origin: origin.value,
     description: description.value,
     attribute: attribute.value,
-    main_image: 'img url',
-    sub_image1: 'img url',
-    sub_image2: 'img url',
-    sub_image3: 'img url',
-    sub_image4: 'img url',
-    sub_image5: 'img url',
+    main_image: file.files[0],
+    sub_image1: file.files[1],
+    sub_image2: file.files[2],
+    sub_image3: file.files[3],
+    sub_image4: file.files[4],
+    sub_image5: file.files[5],
   }
 
   // 제품 데이터 전송
-  fetch('/admin/products', {
+  const token = localStorage.getItem('jwt')
+  fetch(`http://localhost:8081/api/v1/admin/products`, {
     method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: token,
+    },
     body: JSON.stringify(datas),
   })
     .then(response => {
-      if (response.redirected) {
-        window.location.href = response.url
-      }
+      console.log(response)
     })
     .catch(error => console.error('Error:', error))
 
@@ -309,12 +363,18 @@ function handleDeleteProduct(e) {
   if (!e.target.classList.contains('product-delete')) return
   const parent = e.target.parentNode.parentNode
   const id = parent.querySelector('.product-id').innerHTML
-  fetch(`/admin/products/${id}`, {
+  const token = localStorage.getItem('jwt')
+  fetch(`http://localhost:8081/api/v1/admin/products/${id}`, {
     method: 'DELETE',
+    headers: {
+      Authorization: token,
+    },
   })
     .then(response => {
-      if (response.redirected) {
-        window.location.href = response.url
+      if (response.ok) {
+        location.reload()
+      } else {
+        console.error('상품삭제를 실패하였습니다.')
       }
     })
     .catch(error => {
@@ -332,7 +392,6 @@ function handleUpdateSave(e) {
   // 수정 저장
   // 수정된 폼 데이터 수집 및 유효성 검사 후 서버로 PUT 요청
   if (e.target.innerHTML !== '저장') return
-  console.log(1)
   const datas = {
     category: document.querySelector('.category1').value,
     title: document.querySelector('.title1').value,
@@ -351,10 +410,12 @@ function handleUpdateSave(e) {
   }
   const id =
     e.target.parentNode.parentNode.querySelector('.product-id').innerHTML
-  fetch(`/admin/products/${id}`, {
+  const token = localStorage.getItem('jwt')
+  fetch(`http://localhost:8081/api/v1/admin/products/${id}`, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
+      Authorization: token,
     },
     body: JSON.stringify(datas),
   })
